requirements.txt
torch>=2.1.2
lpips
omegaconf
transformers
safetensors
accelerate
imageio[ffmpeg]
PyMCubes
trimesh
megfile
opencv-python
optimum[onnxruntime-gpu]
rembg[gpu,cli]
httpx[socks]
